ENGR00123982: Fix GPU clocking scheme on MX51 & MX53.
Add appropriate secondary clocks to GPU2d and GPU3d clocks.
Also fix the naming scheme for the clocks needed by GPU so
that the same clocks can be used on both MX51 and MX53.
Fixed the round_rate function for many clocks so that the rate
returned is not greater than the maximum rate supported for that clock.

Signed-off-by: Ranjani Vaidyanathan-RA5478 <Ranjani.Vaidyanathan@freescale.com>